Renames reading-materials collection > reads, adds Notifications class, refactors models to favor comp over inheritance and to use more basic object syntax
